requirements_win.txt
jupyter~=1.0.0
pip~=24.0
attrs~=23.2.0
Pillow~=10.0.1
tornado~=6.4
Jinja2~=3.1.2
filelock~=3.12.4
redis~=5.0.1
setuptools~=69.1.1
packaging~=23.2
pyparsing~=3.1.1
torch~=2.0.1
numpy~=1.26.0
tqdm~=4.66.1
opencv-python~=4.9.0.80